+++ internal/helpers/helpers.go
@@ -1,6 +1,10 @@
-package main
+package helpers
 
-import internalTypes "wuzapi/internal/types"
+import (
+	internalTypes "wuzapi/internal/types"
+
+	"github.com/rs/zerolog/log"
+)
 
 func Find(slice []string, val string) bool {
 	for _, item := range slice {
@@ -12,7 +16,7 @@ func Find(slice []string, val string) bool {
 }
 
 // Update entry in User map
-func updateUserInfo(values interface{}, field string, value string) interface{} {
+func UpdateUserInfo(values interface{}, field string, value string) interface{} {
 	log.Debug().Str("field", field).Str("value", value).Msg("User info updated")
 	values.(internalTypes.Values).M[field] = value
 	return values
